Add code for simple server
The server.c file contains an example of a simple TCP/IPv4 server that
you can connect to and send data to. We need to later modify this so
that it uses UDP and also write a (UDP) client.

Comments in the code should be descriptive of the way the UNIX sockets
are used.

To try the program out compile it and run it. Then open another terminal
and write:

> telnet 127.0.0.1 7890

The connection should be initiated and you should be able to send
messages from telnet to the program. To exit the session kill the
server.

For RoboCup: we would need to make it easy to specify IP addresses and
ports because they will change from game to game.

Change-Id: I66257b6b1ad14a74de9b110abaa62f93e0dc865c