Make server error messages more readable
New messages read "server: fatal error <old-message>". This makes it
easy to differentiate the message from a message printed by the client.
Arguably, these would be in different terminals, but they could also be
ran in parallel from the same terminal on the same machine, thus
competing for stderr/stdout.

Change-Id: I9b5630055735933c8259149863e2d8240973df0d