Add a client program; make server output human
The client can connect to any server over TCP/IP and then sends it
"Hello, Server!", after which it echoes everything the server sends
back. The server, on the other hand, waits for a client to connect and
sends "Hello, Client!". Then it echoes everything the client sends to
it. If either disconnects, the other reports this.

Change-Id: Ic22283cea3baefb4c71ecca82649a8c1e47b4a62