Add comms interface as per Design Doc
The Communication module implements the design described in the
Communications API design doc. This consists of a class encapsulating
the implementation and exposing the desired interface functions.

The next steps are to:
 - write a wrapper in ROS to integrate in edante
 - write implementation ... this would also use ROS

The concept of how we can have a clean interface and use ROS at the same
time needs to be clarified.

Change-Id: I658ae62d9598009dc3c58dc07611479a763ef635