Extend Comms; Add a test; Add SPLSTDM;
Add getSTDM method to the Communications class which returns a copy of a
local variable. Implement the setters of the class.

Add the SPL standard message class from SPL's official site.
Override the output stream operator of the SPLSTDM for debugging
purposes.

Add a test file for the Communications class to the project in the tests
folder. It tests the setters of the class as well as the overriden
stream operator.

Change the Space and Time classes to structs.

Change-Id: I6cddb19187e97edbb2ff0dc400cdd7ebd54ee48f