Update Networking interface; create test file
Update Networking.h to include a constructor and data fields for its
functions. It currently also includes the number of team mates,
their addresses and ports as constants. Add missing #endif at
the end. And also add return value to receiveStandardMessage and remove
its argument.

Add Networking.cpp, which currently implements the constructor of
the Networking class and the sendStandardMessage function of
Networking.h.

Add SPLStandardMessage.h, copy of SPLStandardMessage from
the comms branch.

Add NetworkingTest.cpp, which tests the Networking interface,
using test data.

Add Makefile for testing to make it easier.
In order to test do the following:
    - open 4 terminals
    - type "nc -u -l <port number here>"
    - open a 5th terminal
    - type "make"
    - type "./test.out"

Change-Id: I0fdd82d84737b8a5d7bcaec098bdd6925e2f4299