Implement receive part in net
Add fields to Networking class to support receiving. This includes a
socket to receive UDP packets on and a structure to keep the address of
the host.

The implementation is using the method developed in the sockets-exp
branch. It is as simplified as possible.

Also, the Makefile is updated in order to ease testing. To test:

> make
> ./test_receive.out &
> ./test_send.out
> fg
> <Ctrl-C>

Change-Id: If1630e6ccdfad39569e94510ae7d27a5a7328822