Read head angles from the camera node
The head angles are read from the camera_node using naoqi and are put
into the shared memory. The head tracker no longer calls the
motion/get_angles service but uses the one from the shared memory.
This set of angles are synchronised as much as possible with the image
acquisition.