Implement a Kalman filter for the ball position
Implement a Kalman filter which is used to track the ball in the
image. The state consists of [x, y, vx, vy, w, h], where (x, y) is
the ball position; (vx, vy) is the ball velocity; (w, h) is the
size of the bounding box of the detected blob.